+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.42.2] - 2023-10-28
+* Support for latest ESP IDF 5.2 dev (master)
+
 ## [0.42.1] - 2023-10-18
 * Fix ambiguous name error #325 - a compilation issue when the NimBLE component is enabled in `esp-idf-sys`
 * Fix compilation issues of the I2S driver for esp32h2 and esp32c2